feat(物理): 实现基础物理引擎和物理组件
- 新增 PhysicsEngine 类，负责物理模拟和碰撞检测
- 添加 PhysicsComponent 组件，用于物体的物理属性和运动
- 在 GameApp 和 Context 中集成物理引擎
- 更新 Scene 和 GameObject 以支持物理引擎
- 在 GameScene 中添加测试对象和物理相关的测试功能
- 修改 update 函数中参数 delta_time 由 double 改为 float

+++ src/engine/component/component.h
@@ -40,7 +40,7 @@ namespace engine::component {
         // 关键循环函数，全部设为保护，只有 GameObject 需要（可以）调用
         virtual void init() {}
         virtual void handleInput(engine::core::Context &) {}
-        virtual void update(double, engine::core::Context &) {}
+        virtual void update(float, engine::core::Context &) {}
         virtual void render(engine::core::Context &) {}
         virtual void clean() {}
     };
